component folder restructure
