+++ client/src/components/home/Center/CenterFriends.jsx
