+++ client/src/components/home/Center/center.scss
