+++ client/src/components/home/ExerciseList/ExerciseListItem.scss
