+++ client/src/components/home/FriendList/FriendListItem.scss
